+++ CHANGELOG.md
@@ -165,3 +165,4 @@
 - feat: add `rn` resource name to workflow and step
 - feat: update workflow.ver to String type
 - feat: add `desc` to store models
+- feat: add dashmap for Package mod